fix: build out crm-platform from the dead-crates archive
Restore the real Customer Data Platform: customer identity/segmentation,
event ingestion pipeline, autonomous agents (lead qualification, churn
prediction, next-best-action), metrics, personalization, and workflow
automation. Fixes two f32/f64 type-mismatch compile errors that made the
archived crate fail to build at all (ChurnPredictionAgent confidence cast,
Customer::health_score ambiguous float literal).

Makes personalization real: recommendations now vary by customer segment
(vip/at_risk/new/high_value get distinct offer sets) instead of always
returning the same hardcoded two offers.

Makes workflow execution real: WorkflowEngine::execute() now evaluates each
step's conditions (segment:<name>, key=value, or "always") against the
customer and marks the execution "skipped" rather than unconditionally
"completed" when conditions aren't met.

Replaces the dead bin/cli.rs (nonexistent Component type) with a working
demo that ingests an event, runs the agent orchestrator, and prints
personalized recommendations.

+++ Omnisystem/src/crates/crm-platform/src/agents.rs
@@ -52,7 +52,7 @@ impl Agent for ChurnPredictionAgent {
 
         AgentDecision {
             action: if risk > 0.7 { AgentAction::Reach } else { AgentAction::Skip },
-            confidence: risk,
+            confidence: risk as f64,
             reasoning: format!("Churn risk: {}", risk),
         }
     }
